backend/pyproject.toml
[project]
name = "calendly-clone-backend"
version = "0.1.0"
description = "Django backend for the Calendly Clone project."
requires-python = ">=3.14"
dependencies = [
    "django==6.0.4",
    "djangorestframework==3.16.1",
    "psycopg[binary]==3.3.3",
    "gunicorn==25.3.0",
    "freezegun==1.5.5",
    "django-timezone-field==7.2.1",
    "tzdata==2026.2",
]

[tool.uv]
package = false
